cloud-functions/demo-slack-slash-command/requirements.txt
Flask==2.1.1
google-cloud-pubsub==2.12.0